LLAMA-373: TV is showing invalid audio gain value in logs
Reason for change:  audio gain value is actually an 8 bit unsigned,
	but the dbus function sends us an int.  So this casts the value
	so that it prints properly.
Test Procedure:     "AudioGainLevel" value is correct in the logs
Risks:              Low

(cherry picked from commit d49c7e76a95edf6cfba35bdc00c0e40dd5d46967)

+++ include/ctrlm_hal_ble.h
@@ -111,7 +111,7 @@ typedef struct {
    char                          name[CTRLM_MAX_PARAM_STR_LEN];
    char                          serial_number[CTRLM_MAX_PARAM_STR_LEN];
    int                           ir_code;
-   int                           audio_gain_level;
+   uint8_t                       audio_gain_level;
    bool                          audio_streaming;
    unsigned int                  touch_mode;
    bool                          touch_mode_settable;
